requirements.txt
pygame
openai
qwen-agent
redis
pymilvus
pymilvus[model]
pymilvus[milvus_lite]
flask
pyyaml